refactor: Consolidate quantity handling logic and remove redundant frontend file
