+++ plugins/woocommerce/client/blocks/assets/js/blocks/add-to-cart-with-options/quantity-selector/block.json
@@ -12,7 +12,6 @@
 		"interactivity": true
 	},
 	"$schema": "https://schemas.wp.org/trunk/block.json",
-	"viewScriptModule": "woocommerce/add-to-cart-with-options-quantity-selector",
 	"style": "file:../woocommerce/add-to-cart-with-options-quantity-selector-style.css",
 	"editorStyle": "file:../woocommerce/add-to-cart-with-options-quantity-selector-editor.css"
 }
